hough

+++ EBV Projekt/EBVProjekt/sobel.cc
@@ -56,11 +56,11 @@ void sobel (Mat_<ushort>& dstImg, const Mat_<uchar>& srcImg)
 
 #define SOBEL1PXFAST(p, pSrc, sys)\
 {\
-	int sX = (-(int)pSrc[-sys - 1] - ((int)pSrc[-1] << 1) - (int)pSrc[+sys - 1]\
-			 + (int)pSrc[-sys + 1] + ((int)pSrc[1]  << 1) + (int)pSrc[+sys + 1] + 3) >> 3;\
+	int sX = (-pSrc[-sys - 1] - (pSrc[-1] << 1) - pSrc[+sys - 1]\
+			 + pSrc[-sys + 1] + (pSrc[1]  << 1) + pSrc[+sys + 1] + 3) >> 3;\
 	\
-	int sY = (-(int)pSrc[-sys - 1] - ((int)pSrc[-sys] << 1) - (int)pSrc[-sys + 1]\
-			 + (int)pSrc[+sys - 1] + ((int)pSrc[+sys] << 1) + (int)pSrc[+sys + 1] + 3) >> 3;\
+	int sY = (-pSrc[-sys - 1] - (pSrc[-sys] << 1) - pSrc[-sys + 1]\
+			 + pSrc[+sys - 1] + (pSrc[+sys] << 1) + pSrc[+sys + 1] + 3) >> 3;\
 	\
 	*p = sobelCode(sX, sY);\
 }
